Complete test_stress_dir
In the previous version, It couldn't pass the stress_dir test. To solve this,
I checked sfs_touch and sfs_mkdir. In both cases, number of files / dentry-
perblock was is not enough to control every exception. There was a problem if
number of files % dentryperblock is 0. It means the function have to use new
direct pointer but that pointer haven't assigned yet.