Complete cpin and cpout
While coding cpin, I experienced lots of segfault. It was because I just forget to zero-fill.
Make sure that zero-fill the second-level inode pointer entry just after disk_read it!
And, don't forget to reject the cpin command if a file's size exceeds sfs' maximum filesize.
Not copy as much as possible!
By the way, implementing cpout was really simple because there was less exception handling.